refactor filesystem utilities to (hopefully) support both CI and local testing
